perf(client): defer conversation autonomous effects
